Extract shared analysis job methods to new class
